Remove osquerier dependency from library entirely
